requirements.txt
APScheduler==3.5.3
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.6
django-bootstrap4==0.0.6
django-classy-tags==0.8.0
django-cms==3.5.2
django-filer==1.3.2
django-formtools==2.1
django-modelcluster==4.2
django-mptt==0.8.7
django-polymorphic==2.0.2
django-sekizai==0.10.0
django-taggit==0.23.0
django-treebeard==4.3
djangocms-admin-style==1.2.8
djangocms-attributes-field==0.3.0
djangocms-column==1.8.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-installer==1.0.1
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.4
djangorestframework==3.8.2
draftjs-exporter==2.1.2
easy-thumbnails==2.5
gevent==1.3.6
greenlet==0.4.15
gunicorn==19.9.0
html5lib==0.9999999
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
lxml==4.2.3
MarkupSafe==1.0
mysqlclient==1.3.13
parsel==1.5.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pytz==2018.4
queuelib==1.5.0
requests==2.19.1
Scrapy==1.5.1
scrapy-djangoitem==1.1.1
selenium==3.14.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==18.7.0
tzlocal==1.5.1
Unidecode==0.4.21
urllib3==1.23
uWSGI==2.0.17.1
w3lib==1.19.0
wagtail==2.2.2
wcwidth==0.1.7
Werkzeug==0.14.1
Willow==1.1
wrapt==1.10.11
zope.interface==4.5.0
